+++ map/autoware_map_loader/src/lanelet2_map_loader/lanelet2_map_loader_node.hpp
@@ -15,8 +15,8 @@
 #ifndef AUTOWARE__MAP_LOADER__LANELET2_MAP_LOADER_NODE_HPP_
 #define AUTOWARE__MAP_LOADER__LANELET2_MAP_LOADER_NODE_HPP_
 
-#include "map_loader/lanelet2_differential_loader_module.hpp"
-#include "map_loader/utils.hpp"
+#include "lanelet2_differential_loader_module.hpp"
+#include "utils.hpp"
 
 #include <autoware/component_interface_specs_universe/map.hpp>
 #include <autoware/component_interface_utils/rclcpp.hpp>
